backend/requirements.txt
# Lightweight dependencies for HuggingFace Spaces
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
# Minimal ML dependencies
tensorflow==2.15.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
pydantic==2.5.0
huggingface_hub==0.19.4
requests==2.31.0
